Replaced existing RealtimeBox with RealtimeBoxBestEffort, merged the tests into a single file and adapted the naming of the existing tests for the RealtimeBox
