+++ CMakeLists.txt
@@ -46,9 +46,6 @@ if(BUILD_TESTING)
   ament_add_gmock(realtime_box_tests test/realtime_box_tests.cpp)
   target_link_libraries(realtime_box_tests realtime_tools)
 
-  ament_add_gmock(realtime_box_best_effort_tests test/realtime_box_best_effort_tests.cpp)
-  target_link_libraries(realtime_box_best_effort_tests realtime_tools)
-
   ament_add_gmock(realtime_buffer_tests test/realtime_buffer_tests.cpp)
   target_link_libraries(realtime_buffer_tests realtime_tools)
